[UPD] added static variable for product id, fixed price variable type

+++ Models/CatProduct.php
@@ -9,7 +9,7 @@ class CatProduct extends Product
 
     protected string $category = "Cat";
 
-    public function __construct(string $image, string $name, string $description, int $price, int $quantity)
+    public function __construct(string $image, string $name, string $description, float $price, int $quantity)
     {
         parent::__construct($image, $name, $description, $price, $quantity);
         $this->setAvailability($quantity);
