+++ Models/DogProduct.php
@@ -9,7 +9,7 @@ class DogProduct extends Product
 
     protected string $category = "Dog";
 
-    public function __construct(string $image, string $name, string $description, int $price, int $quantity)
+    public function __construct(string $image, string $name, string $description, float $price, int $quantity)
     {
         parent::__construct($image, $name, $description, $price, $quantity);
         $this->setAvailability($quantity);
